lots of null checks

+++ Application/FleetSpeak/app/src/main/java/se/chalmers/fleetspeak/audio/sound/SoundOutputController.java
@@ -76,7 +76,8 @@ public class SoundOutputController implements Runnable {
         if(audioTrack != null){
             audioTrack.release();
         }
-        audioOutputProcessor.terminate();
+        if(audioOutputProcessor != null)
+            audioOutputProcessor.terminate();
     }
